removed old interface unit test, renamed others
