+++ .github/workflows/c.yml
@@ -36,13 +36,8 @@ jobs:
         run: |
           make
           LD_LIBRARY_PATH=../../../rebound/src/ ./rebound
-      - name: Reproduce Holman trajectory (plain interface)
-        working-directory: ./unit_tests/plain_interface/
-        run: |
-          make
-          LD_LIBRARY_PATH=../../../rebound/src/ ./rebound
-      - name: Reproduce Holman trajectory (plain interface, backwards in time)
-        working-directory: ./unit_tests/plain_interface_reverse/
+      - name: Reproduce Holman trajectory (backwards in time)
+        working-directory: ./unit_tests/holman_reverse/
         run: |
           make
           LD_LIBRARY_PATH=../../../rebound/src/ ./rebound
